Add GitHub action workflow to run linters
This commit adds a workflow running the action-ros2-lint GitHub action.

It installs ROS 2 Eloquent linters from OSRF Debian repositories, and
run the linters on the specified packages.

The interest of having a separate workflow for linting (compared to
buildling the package from source and running colcon build and colcon
test), is to provide quick feedback to contributors proposing pull
requests (<1 minute vs 10+ minutes for a complete build and run of the test
suite).

The linters also are running in parallel so that a failure of a linter
does not prevent the other linters from running.

While there may be a small difference between the ament_* linters
provided as Debian packages, and the one on master, we estimate that it
is an acceptable trade-off given the overhead of compiling ament_lint
from source. If this proves problematic, we will add the possibility for
action-ros2-lint to build ament_lint from source.

Signed-off-by: Thomas Moulard <tmoulard@amazon.com>