Fix and clarify logic in test_play filter test
The test was allowing the filter to be incorrect by not checking for the correct number of messages, while setting the filter in a way that was overridden by play_options_.

This doesn't fix the API confusion - I don't have time for that right now. The `Reader.set_filter` call will always be overridden in `rosbag2_transport.play`, which uses `PlayOptions.topics_to_filter`.

Signed-off-by: Emerson Knapp <eknapp@amazon.com>