Remove JumpHandler copy-implementation from PlayerClock/TimeControllerClock (#935)
* Remove JumpHandler copy-implementation from PlayerClock/TimeControllerClock

This API was originally envisioned as a way to handle the `seek`/`jump` operation, but that feature ended up being implemented inline in the `Player`. This is not used now, and I do not anticipate using it in the future. Removing as housekeeping - if it is needed later it can always be re-added from history.

Signed-off-by: Emerson Knapp <emerson.b.knapp@gmail.com>