Enable sanitizers only if code actually can run (#572) (#953)
Uses check_cxx_source_runs instead of check_cxx_source_compiles so that
sanitizers are disabled in an environment that they cannot run in. In
particular this disables the sanitizers on the buildfarm because they're
run in docker containers without the ptrace capability. This should fix
the test test_ros2_message failing on the ROS buildfarm.

Signed-off-by: Shane Loretz <sloretz@openrobotics.org>
Signed-off-by: Shane Loretz <sloretz@osrfoundation.org>

Co-authored-by: Shane Loretz <sloretz@openrobotics.org>