Change the topic names in test_record.cpp (#988)
We've observed the 'test_record' series of tests sometimes
succeeding and sometimes failing on https://ci.ros2.org, particularly
in Windows builds.  When it does fail, it shows error messages
like this:

[ERROR] [1649769476.588754000] [ROSBAG2_TRANSPORT]: Topic '/chatter' has more than one type associated. Only topics with one type are supported

Given that we run tests sequentially (i.e. not in parallel), that
really shouldn't be the case.

However, we have noticed in the past that Windows can sometimes
be extremely slow to cleanup processes from previous tests.
Since many of our tests use a topic name of 'chatter', that is
likely what is happening in this case.  This change works around
this problem by using an alternate name for the topics.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>