Make peek_next_message_from_queue return a SharedPtr.
It is unsafe to return a pointer to a SharedPtr here,
and then store it; if another thread comes along and
pops the queue, then we are now holding a pointer to an
object that has been freed.  Instead, dereference the
pointer so we get a SharedPtr (and a reference) immediately,
which ensures that the object won't be destroyed if we
race with another thread.

While I was in here, did a small refactoring of the method
to remove some duplicated code.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>