Make unpublished topics unrecorded by default (#968) (#1008)
* Make unpublished topics unrecorded by default

Topics with no publishers do not offer any QoS profiles. This causes the
recorder to subscribe with a default QoS, which is often disadvantageous.

This change makes subscribing to unpublished topics optional, off by
default. Once a publisher is discovered for a topic, discovery picks it
up and adds it to the bag.

This addresses #967

Signed-off-by: Sean Kelly <skelly@seegrid.com>

* Feedback from review

Signed-off-by: Sean Kelly <skelly@seegrid.com>

* Additional review feedback

Signed-off-by: Sean Kelly <skelly@seegrid.com>

* Narrow scope of tests and make more deterministic

Signed-off-by: Sean Kelly <skelly@seegrid.com>

* Fix test expectation for unpublished topics

Signed-off-by: Sean Kelly <skelly@seegrid.com>

* Add ROSBAG2_TRANSPORT_EXPORT for get_requested_or_available_topics()

This is needed to try to fix Windows build

Co-authored-by: Sean Kelly <skelly@seegrid.com>
Co-authored-by: Michael Orlov <morlovmr@gmail.com>

Co-authored-by: Sean Kelly <skelly@seegrid.com>
Co-authored-by: Michael Orlov <morlovmr@gmail.com>