Refactor rosbag2_transport::Player to allow two modes of publishing /clock updates.
An additional mode of operation has been added to the Player class to allow for /clock updates. When this mode is active, /clock updates are triggered whenever messages are replayed rather than at a fixed rate. Optionally, a list of triggering topics can be set so that only a subset of replayed messages will trigger the /clock update.

This mode is most useful when replaying applications which do some sanity checking or correlation of message data to system timestamps. If the application does not need the sim time to be updated at a consistent rate, this mode can substantially reduce the overhead of having sim time enabled in rosbag2.

Signed-off-by: Kyle Marcey <kyle.marcey@apex.ai>