zstd_vendor: Fix Findzstd.cmake to properly respect version
Even with the Findzstd.cmake being found, the zstd_VERSION is not
set in Findzstd.cmake, so that the version check always failes even
though a recent version of zstd is available in the system.

The version check is implemented in the Findzstd.cmake and done
with find_package(<library> <version>) and
find_package_handle_standard_args.

Since we cannot use pkg-config for some Windows issues, the parsing of
the version is done by looking for the string in zstd.h.

This unfortunately makes this patch a little verbose.

Tested with ubuntu, gentoo, yocto.

Signed-off-by: Matthias Schoepfer <m.schoepfer@rethinkrobotics.com>