Update rosbag2 to C++17. (#1238)
* Update rosbag2 to C++17.

The main reason to do this is so that we can compile rosbag2
with the clang static analyzer.  As of clang++-14 (what is in
Ubuntu 22.04), the default still seems to be C++14, so we need
to specify C++17 so that new things in the rclcpp headers work
properly.

Further, due to reasons I don't fully understand, I needed to
set CMAKE_CXX_STANDARD_REQUIRED in order for clang to really use
that version.  So set this as well.

We end up skipping rosbag2_storage_sqlite3, since Windows doesn't
like it when that is changed to C++17.  We'll have to revisit this another
time.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>