Add python3-yaml as a dependency (#1490)
Verbs such as `ros2 bag play` import the Python yaml package. This
package is probably present in most ROS installations, but in some
minimalist ones (e.g. built with the Nix package manager), it can be
missing.

Here, we add the dependency to package.xml to ensure that the Python
yaml package is always available.

Signed-off-by: Michal Sojka <michal.sojka@cvut.cz>