Make some changes for newer versions of uncrustify.
Uncrustify 0.78 is the new version in Ubuntu 24.04.
It has some small differences in indentation compared
to Uncrustify 0.72, which is what is in Ubuntu 22.04.

In this PR, we make the changes to match what Uncrustify
0.78 wants.  In most cases, I think this is a "better"
indentation.

We also add in strategic use of "*INDENT-OFF*" comments,
which instructs Uncrustify to not indent this particular
section of code.  That is so that this will still pass
the linters on Uncrustify 0.72 while Rolling is still
on that version.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>