Suppress mcap warnings.
When building humble with Conda, we get lots of warnings
about dll-import problems.  However, we know that these
are safe to ignore for mcap, as this library is strictly
used internally and we suppress this on Rolling.  Do
the same suppression here.

(to be perfectly frank, I don't understand why this is a
problem with conda and not with the current setup, but this
is an easy enough solution)

Signed-off-by: Chris Lalancette <clalancette@gmail.com>